Chore : 컴포넌트 별로 파일 분리
