+++ src/components/Contents.js
@@ -1,15 +1,6 @@
-import Cards from "./Cards";
-import search from "../assets/Search.svg";
 import "../css/Contents.css";
-
-function SearchBar() {
-  return (
-    <div className="search-bar-box">
-      <img src={search} alt="search-icon" />
-      <input className="search-bar" placeholder=" 링크를 검색해 보세요" />
-    </div>
-  );
-}
+import Cards from "./Cards";
+import SearchBar from "./SearchBar";
 
 function Contents({ folderData }) {
   return (
