+++ src/components/Footer.js
@@ -1,7 +1,7 @@
 import facebook from "../assets/facebook.svg";
+import instagram from "../assets/instagram.svg";
 import twitter from "../assets/twitter.svg";
 import youtube from "../assets/youtube.svg";
-import instagram from "../assets/instagram.svg";
 import "../css/Footer.css";
 
 function Footer() {
